% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correct_Nuts_xlsx.R
\name{Correct_Nuts_xlsx}
\alias{Correct_Nuts_xlsx}
\title{Find NUTS of a register file with xlsx extension using 2016 eurostat NUTS.}
\usage{
Correct_Nuts_xlsx(name_of_the_file, country, csv = F, format = "xlsm")
}
\arguments{
\item{name_of_the_file}{It is the name of the register file without the format extension. In order to avoid errors,
it is highly recommended to set the folder that contains the file as the working directory of R.}

\item{country}{character (length == 2), it specifies the country code Aplha-2 of the country.}

\item{csv}{logical (T/F). It specifies if a csv file (clone of the location worksheet but updated),
has to be created in the working directory or not.}
}
\value{
The function returns a dataframe (or a csv file) that contains the location worksheet updated with
all the available NUTS-3 code.
}
\description{
Using NUTS3-Zipcode correspondence tables of eurostat, the function import the pre-specified xlsm file
and insert all the 2016 eurostat NUTS basing on the ZIPCODE.
}
\examples{
## Not Run:
Belgium <- Correct_Nuts_xlsm("register_ES_final", "es")


 Correct_Nuts_xlsm("register_ES_final", "es", T)


## End (Not run)
}
\references{
\href{https://it.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2}, \href{http://ec.europa.eu/eurostat/tercet/flatfiles.do}{Eurostat: Postcode and Nuts}
}
\author{
Massimo Caprari
}
